update backend and web: 1. add retail function

+++ Backend/src/main/java/com/dql/retailmanager/entity/Purchase.java
@@ -42,7 +42,7 @@ public class Purchase implements Serializable {
     /**
      * @mbg.generated 2021-04-27 14:58:45
      */
-    private Integer totalMoney;
+    private Double totalMoney;
 
     private Integer number;
